+++ backend/stout/src/main/java/edu/mtech/stout/resources/CourseResource.java
@@ -1,8 +1,12 @@
 package edu.mtech.stout.resources;
 
+import edu.mtech.stout.api.QueryBySelector;
 import edu.mtech.stout.api.Status;
 import edu.mtech.stout.core.Course;
+import edu.mtech.stout.core.User;
 import edu.mtech.stout.db.CourseDAO;
+import edu.mtech.stout.db.ProgramDAO;
+import io.dropwizard.auth.Auth;
 import io.dropwizard.hibernate.UnitOfWork;
 import io.dropwizard.jersey.PATCH;
 import io.dropwizard.jersey.params.LongParam;
@@ -10,22 +14,33 @@ import io.dropwizard.jersey.params.LongParam;
 import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import java.util.Optional;
 
 @Path("/courses/{courseId}")
 @Produces(MediaType.APPLICATION_JSON)
 public class CourseResource {
 
   CourseDAO dao;
+  QueryBySelector qbs;
 
-  public CourseResource(CourseDAO dao) {
+  public CourseResource(CourseDAO dao, ProgramDAO pDao) {
+    qbs = new QueryBySelector(pDao);
     this.dao = dao;
   }
 
   @GET
   @RolesAllowed({"Admin", "Program Coordinator", "Faculty"})
   @UnitOfWork
-  public Course getCourse(@PathParam("courseId") LongParam courseId) {
-    return findSafely(courseId.get());
+  public Course getCourse(@Auth User user, @PathParam("courseId") LongParam courseId) {
+    Optional<Course> c = dao.findById(courseId.get());
+    if(c.isPresent()){
+      if(qbs.queryByProgramId(user,c.get().getProgramId())){
+        return findSafely(courseId.get());
+      }
+      throw new NotAuthorizedException("Cannot create a course not in your program");
+    }else{
+      throw new NotFoundException("No courses are available in your program.");
+    }
   }
 
   private Course findSafely(long courseId) {
@@ -35,28 +50,41 @@ public class CourseResource {
   @PATCH
   @RolesAllowed({"Admin", "Program Coordinator", "Faculty"})
   @UnitOfWork
-  public Course updateCourse(@PathParam("courseId") LongParam courseId, Course course) {
-    return dao.update(course);
+  public Course updateCourse(@Auth User user, @PathParam("courseId") LongParam courseId, Course course) {
+      if(qbs.queryByProgramId(user,course.getProgramId())){
+        return dao.update(course);
+      }
+      throw new NotAuthorizedException("Cannot create course not in your program");
+
   }
 
   @DELETE
   @RolesAllowed({"Admin", "Program Coordinator"})
   @UnitOfWork
-  public Status deleteCourse(@PathParam("courseId") LongParam courseId) {
-    Status status = new Status();
-    status.setId(courseId.get().intValue());
-    status.setAction("DELETE");
-    status.setResource("Course");
+  public Status deleteCourse(@Auth User user, @PathParam("courseId") LongParam courseId) {
+    Optional<Course> c = dao.findById(courseId.get());
+    if(c.isPresent()){
+      if(qbs.queryByProgramId(user,c.get().getProgramId())){
+        Status status = new Status();
+        status.setId(courseId.get().intValue());
+        status.setAction("DELETE");
+        status.setResource("Course");
 
-    boolean success = dao.delete(findSafely(courseId.get().intValue()));
+        boolean success = dao.delete(findSafely(courseId.get().intValue()));
 
-    if (success) {
-      status.setMessage("Successfully deleted course");
-      status.setStatus(200);
-    } else {
-      status.setMessage("Error deleting course");
-      status.setStatus(500);
+        if (success) {
+          status.setMessage("Successfully deleted course");
+          status.setStatus(200);
+        } else {
+          status.setMessage("Error deleting course");
+          status.setStatus(500);
+        }
+        return status;
+      }
+      throw new NotAuthorizedException("Cannot delete course not in your program");
+    }else{
+      throw new NotFoundException("No courses are available in your program.");
     }
-    return status;
+
   }
 }
